# Optimize Queue<T>.Enumerator