Related to https://github.com/dotnet/runtime/pull/117328

| Method             Toolchain         | Wraps | Length | Mean       | Ratio | Allocated |
|------------------ ------------------ |------ |------- |-----------:|------:|----------:|
| IntsEnumerable     \main\corerun.exe | False | 3      |  18.187 ns |  1.00 |         - |
| IntsEnumerable     \pr\corerun.exe   | False | 3      |   4.249 ns |  0.23 |         - |
|                                      |       |        |            |       |           |
| IntsQueue          \main\corerun.exe | False | 3      |   9.710 ns |  1.00 |         - |
| IntsQueue          \pr\corerun.exe   | False | 3      |   3.933 ns |  0.41 |         - |
|                                      |       |        |            |       |           |
| StringsEnumerable  \main\corerun.exe | False | 3      |  19.448 ns |  1.00 |         - |
| StringsEnumerable  \pr\corerun.exe   | False | 3      |   4.574 ns |  0.24 |         - |
|                                      |       |        |            |       |           |
| StringsQueue       \main\corerun.exe | False | 3      |   9.564 ns |  1.00 |         - |
| StringsQueue       \pr\corerun.exe   | False | 3      |   3.026 ns |  0.32 |         - |
|                                      |       |        |            |       |           |
| IntsEnumerable     \main\corerun.exe | False | 100    | 212.356 ns |  1.00 |      40 B |
| IntsEnumerable     \pr\corerun.exe   | False | 100    | 119.884 ns |  0.56 |         - |
|                                      |       |        |            |       |           |
| IntsQueue          \main\corerun.exe | False | 100    | 260.983 ns |  1.00 |         - |
| IntsQueue          \pr\corerun.exe   | False | 100    |  76.163 ns |  0.29 |         - |
|                                      |       |        |            |       |           |
| StringsEnumerable  \main\corerun.exe | False | 100    | 363.271 ns |  1.00 |      40 B |
| StringsEnumerable  \pr\corerun.exe   | False | 100    | 126.384 ns |  0.35 |         - |
|                                      |       |        |            |       |           |
| StringsQueue       \main\corerun.exe | False | 100    | 250.484 ns |  1.00 |         - |
| StringsQueue       \pr\corerun.exe   | False | 100    |  76.951 ns |  0.31 |         - |
|                                      |       |        |            |       |           |
| IntsEnumerable     \main\corerun.exe | True  | 3      |  16.776 ns |  1.00 |      40 B |
| IntsEnumerable     \pr\corerun.exe   | True  | 3      |   5.312 ns |  0.32 |         - |
|                                      |       |        |            |       |           |
| IntsQueue          \main\corerun.exe | True  | 3      |   9.567 ns |  1.00 |         - |
| IntsQueue          \pr\corerun.exe   | True  | 3      |   2.911 ns |  0.30 |         - |
|                                      |       |        |            |       |           |
| StringsEnumerable  \main\corerun.exe | True  | 3      |  20.782 ns |  1.00 |         - |
| StringsEnumerable  \pr\corerun.exe   | True  | 3      |   3.919 ns |  0.19 |         - |
|                                      |       |        |            |       |           |
| StringsQueue       \main\corerun.exe | True  | 3      |  14.296 ns |  1.00 |         - |
| StringsQueue       \pr\corerun.exe   | True  | 3      |   3.835 ns |  0.27 |         - |
|                                      |       |        |            |       |           |
| IntsEnumerable     \main\corerun.exe | True  | 100    | 271.858 ns |  1.00 |      40 B |
| IntsEnumerable     \pr\corerun.exe   | True  | 100    |  99.918 ns |  0.37 |         - |
|                                      |       |        |            |       |           |
| IntsQueue          \main\corerun.exe | True  | 100    | 245.360 ns |  1.00 |         - |
| IntsQueue          \pr\corerun.exe   | True  | 100    |  82.029 ns |  0.33 |         - |
|                                      |       |        |            |       |           |
| StringsEnumerable  \main\corerun.exe | True  | 100    | 361.511 ns |  1.00 |      40 B |
| StringsEnumerable  \pr\corerun.exe   | True  | 100    | 101.535 ns |  0.28 |         - |
|                                      |       |        |            |       |           |
| StringsQueue       \main\corerun.exe | True  | 100    | 253.267 ns |  1.00 |         - |
| StringsQueue       \pr\corerun.exe   | True  | 100    |  83.072 ns |  0.33 |         - |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkSwitcher.FromAssembly(typeof(Bench).Assembly).Run(args);

[MemoryDiagnoser(false)]
public class Bench
{
    private IEnumerable<int> _intsEnumerable;
    private Queue<int> _intsQueue;
    private IEnumerable<string> _stringsEnumerable;
    private Queue<string> _stringsQueue;

    [Params(false, true)]
    public bool Wraps { get; set; }

    [Params(3, 100)]
    public int Length { get; set; }

    [GlobalSetup]
    public void Setup()
    {
        _intsQueue = new Queue<int>(Length);
        _stringsQueue = new Queue<string>(Length);

        for (int i = 0; i < Length; i++)
        {
            _intsQueue.Enqueue(i);
            _stringsQueue.Enqueue(i.ToString());
        }

        if (Wraps)
        {
            for (int i = 0; i < Length - 1; i++)
            {
                _intsQueue.Dequeue();
                _intsQueue.Enqueue(i + Length);

                _stringsQueue.Dequeue();
                _stringsQueue.Enqueue((i + Length).ToString());
            }
        }

        _intsEnumerable = _intsQueue;
        _stringsEnumerable = _stringsQueue;
    }

    [Benchmark]
    public int IntsEnumerable()
    {
        int sum = 0;
        foreach (var i in _intsEnumerable) sum += i;
        return sum;
    }

    [Benchmark]
    public int IntsQueue()
    {
        int sum = 0;
        foreach (var i in _intsQueue) sum += i;
        return sum;
    }

    [Benchmark]
    public int StringsEnumerable()
    {
        int sum = 0;
        foreach (var s in _stringsEnumerable) sum += s.Length;
        return sum;
    }

    [Benchmark]
    public int StringsQueue()
    {
        int sum = 0;
        foreach (var s in _stringsQueue) sum += s.Length;
        return sum;
    }
}
```